xml2database/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
black==24.8.0
click==8.1.7
Django==5.1.1
djangorestframework==3.15.2
iniconfig==2.0.0
lxml==5.3.0
mypy-extensions==1.0.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.3.3
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.1
